==> codes/category/12.0-product-types-multi-label.ipynb <==
Multi-labeling task: preprocessing data into embeddings 

1. use fasttext  ["Enriching Word Vectors with Subword Information"](https://arxiv.org/abs/1607.04606)
2. build an nn (with bce loss)

## Drive Mounting, directory paths

# !pip -q install bertopic

from google.colab import drive
drive.mount('/content/drive')

maindir = '/content/drive/MyDrive/FinTech-final-project'
datadir = f'{maindir}/data'
spmdir = f'{maindir}/spm'
cat_df_path = f'{maindir}/東吳課程_發票資料集/品類資料集/cat_train_v2.csv'

## Tokenization: SentencePiece Tokenizer 
trained using all names in category dataset 

not very effective 

!pip install gensim

# 斷詞 
!pip -q install sentencepiece

import sentencepiece as spm
sp = spm.SentencePieceProcessor()
sp.Load(f'{spmdir}/spm_allcats.model')

text = 'IM MEME我愛澄露潤唇膏002'
text = 'KissMe花漾美姬瞬翹自然捲纖長防水睫毛膏-02玫瑰棕6g'
text = '龜丸大豆本釀醬油500m'
sp.EncodeAsPieces(text) # not very effective? check why if there's time 

## Tokenization: ckip-transformers 


!pip install -q -U ckip-transformers

from ckip_transformers.nlp import CkipWordSegmenter
ws_driver = CkipWordSegmenter(level=3, device = 0) # device 0 specifies GPU 

import pandas as pd 
catdf = pd.read_csv(cat_df_path)
name_texts = catdf['name'].to_list()
ws = ws_driver(name_texts)

## FastText 
other options: sentenceBERT

# toy-example 
from gensim.models import FastText
from gensim.test.utils import common_texts  # some example sentences
print(common_texts[0])
print(len(common_texts))
# model = FastText(size=4, window=3, min_count=1)  # instantiate # vector_size or size??
# model.build_vocab(sentences=common_texts)
# model.train(sentences=common_texts, total_examples=len(common_texts), epochs=10)  # train

ws = [[x.replace(' ', '') for x in sublist] for sublist in ws]

## cleaning spaces 

# ws: word-segmented (tokenized) texts 
# embedding_dim: 200 
# context_window: 5 
model = FastText(size=200, window=5, min_count=1)  # instantiate # vector_size or size??
model.build_vocab(sentences=ws)
model.train(sentences=ws, total_examples=len(ws), epochs=10)  # train

vocab = model.wv.vocab

len(vocab)

list(vocab.keys())[-30:]

len(ws) == len(catdf) # 約90000筆

model.most_similar('舒點眼液')

import os
fasttextdir = f'{maindir}/models'
os.makedirs(fasttextdir, exist_ok = True)
fasttext_path = f'{fasttextdir}/fasttext.model'
model.save(fasttext_path)
