==> codes/category/15.0_multi_label_simple_nn.ipynb <==
## Fintech Final Project
## Category: multi-label classification 
Author: Nana

Date: Created at 2022.4.14


from google.colab import drive
drive.mount('/content/drive')
maindir = '/content/drive/MyDrive/FinTech-final-project'
datadir = f'{maindir}/data'
spmdir = f'{maindir}/spm'
modeldir = f'{maindir}/models'
cat_df_path = f'{maindir}/東吳課程_發票資料集/品類資料集/cat_train_v2.csv'

## Task: Classifying fasttext embeddings
2 linear layers 

import joblib, sys

# required for skmultilearn (can skip if not using the package)
# # follow the instructions on https://github.com/thunlp/OpenNE

## Preparing Data (X, y) 

http://scikit.ml/multilabeldnn.html#Multi-label-deep-learning-with-scikit-multilearn



fmodelpath = f'{modeldir}/fasttext.model'
ypath = f'{datadir}/category/category_labels_v2.pkl'

import joblib 
from collections import defaultdict
LabelsList = joblib.load(ypath)

CatsDict = defaultdict(list)
for x in LabelsList:
  # print(x.keys())
  for cat in x['product']:
    # print(x['name'])
    CatsDict[cat].append(x['name'])

import random
print('# of categories:', len(CatsDict.keys()))
for k, v in CatsDict.items():
  print(f"Category: {k}, Length: {len(v)}, \nExamples:{random.sample(v, min(10, len(v)))}")
  print('---------------')
  break

import json
with open(f"{datadir}/category/category_dictionary.json", "w") as outfile:
    json.dump(CatsDict, outfile, ensure_ascii = False, indent = 4)

# average the token vectors to form X's feature vector
# use the binary tuple as label
from gensim.models import FastText
# should save keyedvectors only if there's no further training 
model = FastText.load(fmodelpath) 

## segmented, should have saved the segmented results for this
# !pip install -q -U ckip-transformers
# from ckip_transformers.nlp import CkipWordSegmenter
# ws_driver = CkipWordSegmenter(level=3, device = 0)
# name_texts = [x['name'] for x in LabelsList]
# ws = ws_driver(name_texts) !nvidia-smi # k80要20分鐘

import numpy as np
# parameters
emb_dim = 200 # (see notebook 12.0)
Xlen = len(LabelsList)
nclass = len(LabelsList[0]['labels'])

# make X(Emb matrix), y (label matrix)
Emb = np.zeros((Xlen, emb_dim))
Labels = np.zeros((Xlen, nclass))

def get_avg_embeddings(tokenlist):
  # if no token has embedding available, yields np.zeros(emb_dim)
  return np.mean([model.wv[tok] if tok in model.wv else np.zeros(emb_dim) for tok in tokenlist], axis = 0)

def get_weighted_embeddings(tokenlist):
  # ref: https://stackoverflow.com/questions/29330792/weighted-averaging-a-list
  rate = np.ones(len(tokenlist))
  mid = int(len(rate)/2)
  # category name通常出現在產品名稱後半部
  rate = np.append(rate[:mid], (rate[mid:]+1))
  # print(rate)
  X = [model.wv[tok] if tok in model.wv else np.zeros(emb_dim) for tok in tokenlist]
  return np.average(X, 
                    axis = 0,
                    weights=rate)

for i in range(Xlen):
    Emb[i,:] = get_weighted_embeddings(LabelsList[i]['seg_name'])
    Labels[i,:] = LabelsList[i]['labels']

get_weighted_embeddings(['黑人','牙膏'])

print(model.wv['黑人'])
print(model.wv['牙膏'])

## Training
stack a simple few-layer neural network 

!pip install -q -U torch

# seeding; normally not working in colab
import torch
from torch import nn
import torch.nn.functional as F
def seeding(myseed):
  torch.manual_seed(myseed)
  torch.cuda.manual_seed(myseed)
  torch.cuda.manual_seed_all(myseed)
  np.random.seed(myseed)
  torch.backends.cudnn.benchmark = False
  torch.backends.cudnn.deterministic = True

# hyperparameters
MYSEED = 42
TESTRATIO = 0.2 
DROPOUT = 0.5
BATCHSIZE = 200 
## Category數：nclass (217) fixed value
seeding(MYSEED)

from sklearn.model_selection import train_test_split
# 一些分工：做k-fold、
X_train, X_test, y_train, y_test = train_test_split(Emb, Labels, 
                                                    test_size=TESTRATIO, 
                                                    random_state= MYSEED)
input_dim = X_train.shape[1]
hd1, hd2 = 128, 256 
output_dim = nclass 
# originally len(np.unique(y_train.rows)) as http://scikit.ml/multilabeldnn.html#Multi-label-deep-learning-with-scikit-multilearn 
# states, but the logic is (if I understand it correctly) very weird

from torch.utils.data import Dataset

class CatDataset(Dataset):
    def __init__(self, X, y=None):
        self.data = torch.from_numpy(X).float()
        if y is not None:
            y = y.astype(np.int)
            self.label = torch.from_numpy(y).float()
        else:
            self.label = None
    def __getitem__(self, idx):
        if self.label is not None:
            return self.data[idx], self.label[idx]
        else:
            return self.data[idx]
    def __len__(self):
        return len(self.data)


from torch.utils.data import DataLoader
batch_size = BATCHSIZE
train_set = CatDataset(X_train, y_train)
val_set = CatDataset(X_test, y_test)
train_loader = DataLoader(train_set, batch_size=batch_size, shuffle=False)
val_loader = DataLoader(val_set, batch_size=batch_size, shuffle=False)

## Architecture(s): Simple Network (linear layers)


class MultiLabelClassifier(nn.Module):
    def __init__(
            self,
            input_dim=input_dim,
            hd1 = hd1,
            hd2 = hd2,
            output_dim=output_dim,
            dropout=DROPOUT,
    ):
        super(MultiLabelClassifier, self).__init__()
        self.dropout = nn.Dropout(dropout)
        # 只有兩層fc layers (shallow)，主要是看fasttext embedding的效果
        self.hidden1 = nn.Linear(input_dim, hd1)
        self.hidden2 = nn.Linear(hd1, hd2)
        self.output = nn.Linear(hd2, output_dim)

    def forward(self, X, **kwargs):
        X = self.hidden1(X)
        X = self.hidden2(X)
        X = F.relu(X)
        X = self.dropout(X)
        # 過softmax拿機率值
        X = F.softmax(self.output(X), dim=-1)
        return X

References: 
https://keras.io/examples/nlp/pretrained_word_embeddings/
https://towardsdatascience.com/

lstm-text-classification-using-pytorch-2c6c657f8fc0

My machine learning assignments

# from skmultilearn.problem_transform import LabelPowerset
# training parameters
THRESHOLD = 0.3
EPOCHS = 50
device = 'cuda' if torch.cuda.is_available() else 'cpu'
learning_rate = 5e-4       # learning rate
net = MultiLabelClassifier().to(device)
criterion = nn.BCELoss() 
weight_decay = 0.1
optimizer = torch.optim.AdamW(net.parameters(), 
                              lr=learning_rate, 
                              weight_decay=weight_decay)

import warnings
warnings.filterwarnings("ignore")

import time, datetime
from datetime import date, datetime
now = datetime.now()
model_prefix = now.strftime("%m%d-%H%M")
MODELPATH = f'{maindir}/multi_label/{model_prefix}.ckpt'
MODELPATH

METHOD =( 'macro', 'weighted')

# Training Loop
from sklearn.metrics import f1_score
best_val_acc = 0.0
best_acc = 0.0

for epoch in range(EPOCHS):
    train_acc, val_acc, val_weighted_acc = 0.0, 0.0, 0.0
    train_loss, val_loss = 0.0,0.0
    # training
    net.train() # set the model to training mode
    for i, data in enumerate(train_loader):
        inputs, labels = data
        # print('l:', labels.shape)
        inputs, labels = inputs.to(device), labels.to(device)
        optimizer.zero_grad() 
        outputs = net(inputs) 
        # print('o:', outputs.shape, 'l:', labels.shape)
        batch_loss = criterion(outputs, labels)
        
        # THRESHOLD: all labels with probabilities higher than it are considered predicted labels 
        # and others are skipped. We are using a threshold value of 0.5.
        pred = np.array(outputs.cpu().detach().numpy() > THRESHOLD, dtype=float)
        labels = labels.cpu().detach().numpy()
        # train_acc += getbatchf1(labels, pred)
        # print(f1_score(labels, pred, average='macro'))
        # compute macro_f1
        # print(labels.shape, pred.shape)
        # print(f1_score(labels, pred, average='macro'))
        batch_loss.backward() 
        optimizer.step() 
        train_loss += batch_loss.item()

    # validation
    if len(val_set) > 0:
        net.eval() # set the model to evaluation mode
        val_preds, val_labels = [], []
        with torch.no_grad():
            for i, data in enumerate(val_loader):
                inputs, labels = data
                inputs, labels = inputs.to(device), labels.to(device)
                outputs =net(inputs)
                batch_loss = criterion(outputs, labels) 
                val_pred = np.array(outputs.cpu().detach().numpy() > THRESHOLD, dtype=float)
                labels = labels.cpu().detach().numpy()
                val_loss += batch_loss.item()
                val_preds.append(val_pred)
                val_labels.append(labels)
            val_labels, val_preds = np.vstack(val_labels), np.vstack(val_preds)
            val_acc = f1_score(val_labels, val_preds, average = METHOD[0])
            val_weighted_acc = f1_score(val_labels, val_preds, average = METHOD[1])
            
            if val_acc > best_val_acc:
              torch.save(net.state_dict(), MODELPATH)
              best_val_acc = val_acc
            print('[{:03d}/{:03d}] Loss: {:3.4f} || VLoss: {:3.4f} | Val macro f1: {:3.4f}, weighted f1: {:3.4f}.'.format(
                epoch + 1, EPOCHS, 
                train_loss/len(train_loader), 
                val_loss/len(val_loader),
                val_acc, 
                val_weighted_acc,
            ))

## **對這結果的檢視與說明：**
https://www.notion.so/4-27-macro-f1-on-multi-label-5c9e0534efe94cc8a526246f66f213e5

from sklearn.metrics import f1_score, classification_report
# multilabel classification
y_true = [[0, 0, 0], [1, 1, 1], [0, 1, 1], [1,0,0]]
y_pred = [[0, 0, 0], [1, 1, 1], [1, 1, 0], [1,0,1]]

sc = 0
for a, b in zip(y_true, y_pred):
  sc+=f1_score(a, b, average = 'macro')
sc/=len(y_true)

y_true, y_pred = np.array(y_true), np.array(y_pred)
f1_score(y_true, y_pred, average='macro') 

net.eval()
val_preds, val_labels = [], []
for i, data in enumerate(val_loader):
      inputs, labels = data
      inputs, labels = inputs.to(device), labels.to(device)
      outputs =net(inputs)
      val_pred = np.array(outputs.cpu().detach().numpy() > THRESHOLD, dtype=float)
      labels = labels.cpu().detach().numpy()
      val_preds.append(val_pred)
      val_labels.append(labels)

val_preds = np.vstack(val_preds)
val_labels = np.vstack(val_labels)
val_labels.shape

val_acc = f1_score(val_labels, val_preds, average = METHOD[0])
val_weighted_acc = f1_score(val_labels, val_preds, average = METHOD[1])
print(f'macro f1: {val_acc:.4f}')
print(f'weighted f1: {val_weighted_acc:.4f}')

cat2idx = joblib.load(f'{datadir}/category/cat2idx.pkl')
labels = sorted(cat2idx.items(), key = lambda x:x[1])
print(labels[:5])
keys = [x[0] for x in labels]
keys[:5]

print(classification_report(val_labels, val_preds, target_names= keys))

len(keys)

## Checking example reference: 
http://scikit.ml/multilabeldnn.html#Multi-label-deep-learning-with-scikit-multilearn
(which doesn't seem to have the logic our task requires)

!pip install arff
from skmultilearn.dataset import load_dataset
X_train, y_train, feature_names, label_names = load_dataset('emotions', 'train')
X_test, y_test, _, _ = load_dataset('emotions', 'test')
