==> codes/category/23.0_category_fasttext_clustering.ipynb <==
## Category: nn之前的做法

Author: Nana

Date: Created at 2022.4.16


# hyperparams
TOPN = 20 
N = 217 # category數量

from google.colab import drive
drive.mount('/content/drive')
maindir = '/content/drive/MyDrive/FinTech-final-project'
datadir = f'{maindir}/data'
spmdir = f'{maindir}/spm'
modeldir = f'{maindir}/models'
cat_df_path = f'{maindir}/東吳課程_發票資料集/品類資料集/cat_train_v2.csv'

from collections import defaultdict
import numpy as np
import pandas as pd
import joblib
import random

## Fixing some bugs in pickled files

catdf = pd.read_csv(cat_df_path)
catD = defaultdict(int)
for k in catdf['product']:
  for x in k.split(';'):
    catD[x] += 1 

N = len(catD)
cat2idx = {} # a mapping that maps category to its corresponding (arbitray) index
for i, k in zip(range(N), catD.keys()):
  cat2idx[k] = i
prod_labels = []
# labs is the v1 category_labels.pkl, now deleted 
# for p, n, l in zip(catdf['product'], catdf['name'], labs):
#   segn = l['seg_name']
#   labels = np.zeros(N)
#   for x in p.split(';'):
#     if len(x.strip()) > 0:
#       catid = cat2idx[x]
#       labels[catid] = 1
#   prod_labels.append({'name':n, 'labels': labels, 'seg_name':segn, 'product': p.split(';')})

# check category number 
all(len(x['labels']) == 217 for x in prod_labels)

## fixing the category number bug
# import joblib
# joblib.dump(prod_labels, f'{datadir}/category/category_labels_v2.pkl')
# joblib.dump(cat2idx, f'{datadir}/category/cat2idx.pkl')

## Loading Fasttext model
`model.wv.most_similar()`

This method computes cosine similarity between a simple mean of the projection weight vectors of the given words and the vectors for each word in the model. The method corresponds to the word-analogy and distance scripts in the original word2vec implementation.

fmodelpath = f'{modeldir}/fasttext.model'
from gensim.models import FastText
model = FastText.load(fmodelpath) 

model.wv.most_similar('人工淚液')

model.wv.most_similar('筆記型電腦')

### Attempt 1: String exact match 
may serve as a baseline 

以FastText內建的`model.wv.most_similar(c)` ∀ category c，存下最相似的前20個(`TOPN`) 詞彙與對應的機率（as recorded in model.wv）作為關鍵詞表（存在`cat2kw`）。隨後對每個product name，string match所有category對應的詞表（手動加入category本身詞彙，權重設定為100，以機率來看= 10000%，幾乎等於一定選它），然後將有match到的詞彙對應的機率加總（cat_hit_probs）。
對每個name，argmax cat_hit_probs最大者作為所屬category。
1. 只選單一category（argmax）。
2. 以下只random sample了10%的資料約9000筆來實驗。

Macro F1 Result: 0.39

!pip install -q -U ckip-transformers
from ckip_transformers.nlp import CkipWordSegmenter

ws_driver = CkipWordSegmenter(level=3) # device 0 specifies GPU 

cat2idx = joblib.load(f'{datadir}/category/cat2idx.pkl')
cat_dict = joblib.load(f'{datadir}/category/category_labels_v2.pkl')

inverted map

idx2cat = {v:k for k, v in cat2idx.items()}

catkeys = list(cat2idx.keys())
cat2ws = {} # mapping: category to its segmented token 
ws = ws_driver(catkeys)
for k, segk in zip(catkeys, ws):
  cat2ws[k] = segk
joblib.dump(cat2ws, f'{datadir}/category/cat2ws.pkl')

# check if the category has an embedding in fasttext model
# if yes, take its most_similar phrase list
# otherwise, try get its tokens' most_similar phrase list (extend)

cat2keywords= defaultdict(list)
for cat in cat2idx.keys():
  if cat in model.wv:
    cat2keywords[cat] = model.wv.most_similar(cat, topn = TOPN)
  else:
    tokens = cat2keywords[cat]
    for token in tokens:
        cat2keywords.extend(model.wv.most_similar(token, topn = TOPN))

list(cat2keywords.items())[:5]

# clean the cat2keywords 
# (remove probabilities and duplicates, add itself)
# clean too-short keywords <- 效果不好
cat2kw = {}
for k, v in cat2keywords.items():
  clean_v = list(set(x[0] for x in v))
  clean_v_probs = {x[0]:x[1] for x in v} # 如果x[0]有重複出現，機率那邊有點麻煩但我們先不處理
  if k not in clean_v:
    clean_v.append(k)
    clean_v_probs[k] = 100 # 100% similar # guarantee to take it
  cat2kw[k] = (clean_v, clean_v_probs)
# cat2kw here is a dictionary of category to its phrase list 

# use regex to find the exact match 
# sampling 0.1 ratio (for speed)
instances_num = len(catdf)
samp_num = int(instances_num/10)
randidx = random.sample(range(instances_num), samp_num)
samples = [cat_dict[i] for i in randidx]

len(samples)

samples[0]

import numpy as np
import re

# phrases: the phrases belonging to the category's most similar 
# re.findall(r"(?=("+'|'.join(phrases)+r"))", x)

# hit_probs

for i, sample in enumerate(samples):
    hit_vector = np.zeros(N)
    name = sample['name']
    for cat, pair in cat2kw.items():
        # v: phraselist belonging to k (category)
        # (?=(喵喵|咩咩))
        ph_list, ph_probs = pair
        pat = re.compile(r'(?:{})'.format('|'.join(map(re.escape, ph_list))))
        found = re.findall(pat, name)
        # the current category hits' probabilites
        cat_hit_probs = sum(ph_probs[x] for x in found)
        catid = cat2idx[cat]
        hit_vector[catid] = cat_hit_probs
    samples[i]['hitsprobs'] = hit_vector 
    

'''randomly pick some instances to examine the result'''
import random
rand = random.randint(0, len(samples))
print('index:', rand)
print(np.where(samples[rand]['hitsprobs'] > 0))
print(np.where(samples[rand]['labels'] ==1))

x = rand
predcat = np.argmax(samples[x]['hitsprobs'])
print(idx2cat[predcat])
samples[x]['hitsprobs'] # 挑argmax，先暫時當one-label　

**First try argmax**

-> simplify the problem into one-label task 
(multiclass, one answer) 

from sklearn.metrics import f1_score
y_true, y_pred = [], []
y_true_labels, y_pred_labels = [], []
for sample in samples:
  predcat= np.argmax(sample['hitsprobs'])
  pred = np.zeros(N); pred[predcat] = 1
  y_pred.append(pred) 
  y_pred_labels.append(idx2cat[predcat])
  y_true_labels.append(sample['product'])
  y_true.append(sample['labels'])

import random
r = random.randint(0, len(samples))
print('index:', r)
print(y_pred_labels[r])
print(y_true_labels[r])

# 10% data; macro f1: 39%
print('10% ratio dataset macro f1:')
print(f1_score(y_true, y_pred,
                  average = 'macro'))
print('10% ratio dataset weighted f1:\n', f1_score(y_true, y_pred,
                  average = 'weighted'))

## Clustering idea: 如何找centroids？
1. Method 1: (by word) ∀ category c，以category name直接找對應的embedding作為c的representative embedding (centroid)。
2. Method 2.1: (by member)
∀ category c，將屬於c下的每個member找切分train:test，以train split內的members找到平均embedding，隨後取他們的average作為該c的centroid。
3. Method 2.2: (by member)
∀ category c，將屬於c下的每個member找切分train:test，以train split內的members找到權重embedding，隨後取他們的average作為該c的centroid。所謂權重embedding，是因為資料內真正和category相關的token通常落在name的中後段，如「視舒坦人工淚液點眼液10ML」（人工淚液）、「697155@霧峰區農會酒莊酒粕香腸600G」（中式香腸），因此將後半段token的權重加乘至W倍（也可想其他加權法）。



Clustering idea: 找到centroids之後？應該有不只一種方法可以嘗試。
### Attempt 2: Clustering 2.2
centroid-finding Method 2.2然後對每個test split內的instance，同樣計算權重embedding，然後用distance function計算其和每個category centroids的距離，切一個距離threshold，將<=的都算1，>的算0。
取10%資料：

Macro F1 Result -> 0.28

name2vec = {}
EMBDIM = 200 # see notebook 12.0 
W = 4

def get_avg_embeddings(tokenlist):
  # if no token has embedding available, yields np.zeros(emb_dim)
  return np.mean([model.wv[tok] if tok in model.wv else np.zeros(EMBDIM) for tok in tokenlist], axis = 0)

def get_weighted_embeddings(tokenlist):
  # ref: https://stackoverflow.com/questions/29330792/weighted-averaging-a-list
  rate = np.ones(len(tokenlist))
  mid = int(len(rate)/2)
  # category name通常出現在產品名稱後半部
  rate = np.append(rate[:mid], (rate[mid:]+(W-1)))
  X = [model.wv[tok] if tok in model.wv else np.zeros(EMBDIM) for tok in tokenlist]
  return np.average(X, 
                    axis = 0,
                    weights=rate)

# step 1. compute weighted embeddings and save into cat_dict
'''
for i in range(len(cat_dict)):
  instance = cat_dict[i]
  toklist = instance['seg_name']
  wemb = get_weighted_embeddings(toklist)
  cat_dict[i]['weighted_emb'] = wemb 
  if not np.any(wemb): # if it is a zero vector
    print(f"{instance['name']} has no token embedding")
'''

# record category lists for each instance into cat_dict
'''
for i in range(len(cat_dict)):
   instance = cat_dict[i]
   cats = catdf.iloc[i]['product']
   cats = cats.split(';')
   cat_dict[i]['categories'] = cats
'''

# save back into cat_dict
'''
joblib.dump(cat_dict, f'{datadir}/category/category_labels_v2.pkl')
'''

TESTSIZE = 0.1
SEED = 42

# step 2. train-test split
from sklearn.model_selection import train_test_split
X = [(x['weighted_emb'], x['name']) for x in cat_dict]
y = [x['labels'] for x in cat_dict]
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size = TESTSIZE, random_state=SEED)
# 感覺要確保一下stratify

# step 3. compute centroids
member_count = np.zeros((217,))
centroids = np.zeros((217, 200))
for (emb,_), labels in zip(X_train, y_train):
    indices = np.nonzero(labels)[0].tolist()
    for catid in indices:
      member_count[catid] += 1
      centroids[catid, :] += emb
# averaging 
for catid in range(N):
  try:
    centroids[catid, :] = centroids[catid, :]/member_count[catid]
  except: 
    print(f'Category idx2cat[catid] has no training member.')

# naked eye看stratification（至少每類要train set裡出現一次）
member_count.astype(int)

# step 4. compute distances (for each test instance, compute for each class)
# m: testing #
# N: category #
# dist_time_comp # time complexity for calling distance function once
# O(m*N*dist_tim_comp)
from numpy import linalg as LA # 2-norm for Euclidean
# LA.norm
TESTNUM = len(X_test)
Dist = np.zeros((TESTNUM, N))
for i in range(TESTNUM):
  x_test, x_name = X_test[i]
  for catid in range(N):
    centroid = centroids[catid, :]
    # print(x_test.shape, centroid.shape)
    dist = abs(LA.norm(x_test - centroid))
    if dist < 1e-2:
      print(idx2cat[id], x_name)
    Dist[i, catid] = dist

# to-do: 
# step 5. STATIC assign member by distance function # the closer the better
gap = 0.1
predictions = []
for i in range(TESTNUM):
    _, x_name = X_test[i]
    real_classes = [idx2cat[e] for e in np.nonzero(y_test[i])[0]]
    labels = np.zeros(N)
    c1_dist, c1 = np.min(Dist[i, :]), np.argmin(Dist[i, :])
    # temporarily mask the argmin class
    Dist[i, c1] = float('inf')
    c2_dist, c2 = np.min(Dist[i, :]), np.argmin(Dist[i, :])
    labels[c1] = 1
    # revise it back 
    Dist[i, c1] = c1_dist
    if i < 10: 
       print(f' * {i+1}. testing instance:', x_name, '\n\ttrue class:',real_classes)
       print(f'\tpredicted class:',idx2cat[c1])
    if abs(c1_dist - c2_dist) < gap:
            labels[c2] = 1
            if i < 10:
                print('\tpredicted class 2:',idx2cat[c2])
    predictions.append(labels)

# step 6. metrics: calculate macro f1
f1_score(y_test, predictions, average = 'macro')

f1_score(y_test, predictions, average = 'weighted')

len(predictions)
