==> codes/category/30.0_train_bert.ipynb <==
# Fine-tuning BERT 
Fintech Project

Author: 楊晴雯 

Date: 2022.5.6 

References:

- https://colab.research.google.com/github/huggingface/notebooks/blob/main/examples/text_classification.ipynb#scrollTo=X6HrpprwIrIz


from google.colab import drive
drive.mount('/content/drive')
maindir = '/content/drive/MyDrive/FinTech-final-project'
datadir = f'{maindir}/data'
spmdir = f'{maindir}/spm'
modeldir = f'{maindir}/models'
cat_df_path = f'{maindir}/東吳課程_發票資料集/品類資料集/cat_train_v2.csv'

!pip -q install datasets

from datasets import Dataset, load_metric
import datasets
import joblib, sys
import numpy as np
import pandas as pd

datapath = f'{datadir}/category/category_labels_v2.pkl'
datas = joblib.load(datapath)

cat2idx = joblib.load(f'{datadir}/category/cat2idx.pkl')
idx2cat = {v:k for k, v in cat2idx.items()}

multi = 0 
Rows = []
for idx in range(len(datas)):
  row = []
  data = datas[idx]
  row.append(idx) # first idx
  row.append(data['name'])
  label = list(np.where(data['labels'] >= 1)[0])
  # 
  if len(label) > 1:
    # print(data['name'], label)
    multi += 1
  # else: assert idx2cat[label[0]] == data['product'][0] <- 全對
  row.append(data['labels'])
  row.append(data['seg_name'])
  row.append(data['product'])
  Rows.append(row)
print('How many multi-labeled data?', multi)
print(f'Ratio of multi-labeled data: {multi/len(datas):.4f}')

#滿煩的，HF🤗沒有提供multi-label的trainer，要自己寫完

## Dataset 

df = pd.DataFrame(Rows, columns=['id', 'name', 'label', 'seg_name', 'label_name'])

df

TESTSIZE = 0.2

dataset = Dataset.from_pandas(df)
trainset, testset = dataset.train_test_split(test_size= TESTSIZE).values()
final_dataset = datasets.DatasetDict({"train":trainset,"test":testset})

final_dataset

## BERT tokenization

!pip -q install transformers

MODELNAME = 'bert-base-chinese'
from transformers import BertTokenizerFast
tokenizer = BertTokenizerFast.from_pretrained(MODELNAME)

def preprocess(examples):
  '''essentailly just tokenization, get input_id, attn_masks, token_type_ids'''
  # print(examples)
  tokenized_name = tokenizer(examples['name'], truncation=True, padding = 'max_length', 
                             max_length = 100)
  examples_ = {k:v for k, v in tokenized_name.items()}
  examples_['name'] = [name for name in examples['name']]
  examples_['label'] = [label for label in examples['label']]
  del examples
  return examples_ 
final_dataset = final_dataset.map(preprocess, batched = True)

from transformers import AutoModelForSequenceClassification
N = len(idx2cat)
print(N)
model = AutoModelForSequenceClassification.from_pretrained(MODELNAME, num_labels = N)

final_dataset

def collate_fn(batch):
    keys = batch[0].keys()
    print(keys)
    # 'id', 'name', 'label', 'seg_name', 'label_name', 'input_ids', 'token_type_ids', 'attention_mask', 'labels'
    to_tensor = ['input_ids', 'token_type_ids', 'attention_mask']
    new_batch = {k:[d[k] for d in batch] for k in keys if k in to_tensor}
    new_batch = tokenizer.pad(
        new_batch,
        padding= "max_length",
        max_length= 100,
        return_tensors="pt",
    )
    label = [d['label'] for d in batch]
    ids = [d['id'] for d in batch]
    new_batch['label'] = torch.tensor(label, dtype=torch.int64)
    new_batch["id"] = torch.tensor(ids, dtype=torch.int64)
    return new_batch

batch_size = 100
lr = 1e-5
epochs = 20
myseed = 1027 # 43 # 1123
gast = 1
wmst = 20
weight_decay = 0.05

from torch.utils.data import DataLoader
train_loader = DataLoader(final_dataset['train'], batch_size = batch_size, collate_fn=collate_fn, shuffle=True)
val_loader = DataLoader(final_dataset['test'], batch_size = batch_size, collate_fn=collate_fn, shuffle = False)

split = 'train'
# print(len(final_dataset[split]['id']))
# print(len(final_dataset[split]['label']))
# print(len(final_dataset[split]['seg_name']))
# print(len(final_dataset[split]['input_ids']))
# print(len(final_dataset[split]['attention_mask']))

## Training loop

import pickle
from dataclasses import dataclass
import torch
from transformers import AutoModelForMultipleChoice, TrainingArguments, Trainer

import random
import torch 
import os
import torch.nn as nn
weight_decay = 0.05
criterion = nn.CrossEntropyLoss()
optimizer = torch.optim.AdamW(model.parameters(),
                              weight_decay = weight_decay, 
                              lr = lr)

!pip -q install wandb

import wandb
from datetime import datetime
import os
wandb.login()
now = datetime.now()
runname = now.strftime("%m%d-%H%M")
model_dir = f'{maindir}/neural_network/{runname}'
os.makedirs(model_dir, exist_ok = True)
print('* Name of the run:', runname)

wandb.init(project = "Fintech Project",
           entity="nana2929", 
           name = runname, 
           config = {
               "batch_size": batch_size,
               "dataset": "麻布數據dataset1.0",
               "learning_rate": lr,
               "epochs": epochs, 
               # "scheduler": SCHETYPE, 
           })

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

colab上gpu記憶體不夠，得挪去server上訓練。

# %%wandb
model.to(device)
for epoch in range(epochs):
  model.train()
  for id, batch in enumerate(train_loader):
      inputs = batch['input_ids'].to(device)
      attnmasks = batch['attention_mask'].to(device)
      labels = batch['label'].to(device)
      logits = model(inputs, attnmasks).logits
      loss = criterion(logits, labels)
      loss.backward()
      torch.nn.utils.clip_grad_norm_(model.parameters(), 1.0)
      optimizer.step()
      print(f'Epoch {epoch+1}| train loss: {loss:.3f}')
    # label length必須要是一致的

# 缺compute_accuracy(), logging...
