==> codes/category/32.0_train_bert_more.ipynb <==
# Fine-tuning BERT-2 [品類分類，多標籤]




Fintech Project

Author: 楊晴雯 

Date: 2022.5.10 



# from google.colab import drive
# drive.mount('/content/drive')
# maindir = '/content/drive/MyDrive/FinTech-final-project'
# datadir = f'{maindir}/data'
# spmdir = f'{maindir}/spm'
# modeldir = f'{maindir}/models'
# cat_df_path = f'{maindir}/東吳課程_發票資料集/品類資料集/cat_train_v2.csv'
datadir = './'

# !pip -q install datasets

import pickle
from dataclasses import dataclass
import torch
import random
import torch 
import os
import torch.nn as nn
import torch.nn.functional as F

from datasets import Dataset, load_metric
import datasets
import joblib, sys
import numpy as np
import pandas as pd

checkpoint_path = './checkpoints/0510-0939/model.ckpt'
datasets_path = './dataset/encoded_dataset_30.0'
encoded_dataset = datasets.load_from_disk(datasets_path)

## Reload BERT checkpoint 

cat2idx = joblib.load(f'{datadir}/category/cat2idx.pkl')
idx2cat = {v:k for k, v in cat2idx.items()}

MODELNAME = 'bert-base-chinese'
from transformers import BertTokenizerFast
tokenizer = BertTokenizerFast.from_pretrained(MODELNAME)
N = len(idx2cat)

model = torch.load(checkpoint_path)

model # A BertForSequenceClassification model based on bert-base-chinese

def collate_fn(batch):
    keys = batch[0].keys()
    # print(keys)
    # 'id', 'name', 'label', 'seg_name', 'label_name', 'input_ids', 'token_type_ids', 'attention_mask', 'labels'
    to_tensor = ['input_ids', 'token_type_ids', 'attention_mask']
    new_batch = {k:[d[k] for d in batch] for k in keys if k in to_tensor}
    new_batch = tokenizer.pad(
        new_batch,
        padding= "max_length",
        max_length= 100,
        return_tensors="pt",
    )
    label = [d['label'] for d in batch]
    ids = [d['id'] for d in batch]
    new_batch['label'] = torch.tensor(label, dtype=torch.float32)
    new_batch["id"] = torch.tensor(ids, dtype=torch.int64)
    return new_batch

extra_epochs = 15
# the loaded model has been trained for 20 epochs already, and reaches: 
# VAL F1 macro: 0.529, weighted: 0.809
# now we want to train more and train with different accuracy computation method
# that is to set the zero_division value to 1 in f1_score() for those missed classes so that the result is more properly 
# evaluated 

batch_size = 200
lr = 1e-5
myseed = 1027 # 43 # 1123
gast = 1
wmst = 20
weight_decay = 0.05

def seeding(myseed):
    torch.manual_seed(myseed)
    torch.cuda.manual_seed(myseed)
    torch.cuda.manual_seed_all(myseed)
    np.random.seed(myseed)
    torch.backends.cudnn.benchmark = False
    torch.backends.cudnn.deterministic = True
seeding(myseed)

## Dataset to Dataloader
Note that the dataset is saved from 30.0 script, otherwise we'll have train test splits mixed up in further training.

from torch.utils.data import DataLoader
train_loader = DataLoader(encoded_dataset['train'], batch_size = batch_size, collate_fn=collate_fn, shuffle=True)
val_loader = DataLoader(encoded_dataset['test'], batch_size = batch_size, collate_fn=collate_fn, shuffle = False)

## Preparation for training loop 

weight_decay = 0.05
criterion = nn.CrossEntropyLoss()
optimizer = torch.optim.AdamW(model.parameters(),
                              weight_decay = weight_decay, 
                              lr = lr)

# !pip -q install wandb

import wandb
from datetime import datetime
import os
wandb.login()
now = datetime.now()
runname = now.strftime("%m%d-%H%M")

wandb.init(project = "Fintech Project",
           entity="nana2929", 
           name = checkpoint_path.split('/')[-2],  #'0510-0939'
           config = {
               "batch_size": batch_size,
               "dataset": "麻布數據dataset1.0",
               "learning_rate": lr,
               "epochs": extra_epochs, 
               "note": 'This is the further training on 0510-0939 checkpoint'
           })

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

!nvidia-smi

## Complete train loop
- With logging, metrics, report to wandb
- colab上gpu記憶體不夠，得挪到server上訓練
- batch_size開1000會太大，可嘗試200~800的範圍

from sklearn.metrics import f1_score
def compute_accuracy(labels, preds, zd = 1):
    '''
    REVISION WARNING:
    Different from 30.0 where we set zero_division to 0 for those classes missed in validation set, 
    we now use 1 for computation of averaged f1. 
    
    F1-score should be computed by the complete validation set! 
    Not by batch, SO BE SURE TO PASS IN two (val_size, nclass) arrays for inputs
    '''
    
    preds = np.where(preds <= 0.5, 0, 1)
    return {'macro f1': f1_score(labels, preds, average = 'macro', zero_division = zd), 
            'weighted f1': f1_score(labels, preds, average = 'weighted', zero_division = zd)}

# y_ = np.array([0, 0.9, 2, 1, 0.1, 0.2]) <- simulate output predictions 
# y_ = np.where(y_ <= 0.5, 0, 1) 
# y = np.array([1, 1, 1, 1, 0, 0]) <- simulate real labels
# f1_score(y, y_, average= 'macro')

# %%wandb
from tqdm import tqdm

model.to(device)
max_val_acc = -1


# the already-trained epoch number
base = 20 

for epoch in range(extra_epochs):
    model.train()
    epoch_loss = {'train': 0.0, 
                  'val': 0.0}
    val_pairs = {'preds': [], 
                'labels': []}
    train_pairs = {'preds': [],
                  'labels': []}
    model.train()
    
    for id, batch in enumerate(tqdm(train_loader)):
        inputs = batch['input_ids'].to(device)
        attnmasks = batch['attention_mask'].to(device)
        labels = batch['label'].to(device)
        logits = model(inputs, attnmasks).logits
        
        preds = F.softmax(logits, dim = 1)
        loss = criterion(preds, labels)
        loss.backward()
        torch.nn.utils.clip_grad_norm_(model.parameters(), 1.0)
        optimizer.step()
        # storing results 
        epoch_loss['train'] += loss.item()
        
        preds = preds.detach().cpu()
        labels = labels.detach().cpu()
        train_pairs['labels'].append(labels)
        train_pairs['preds'].append(preds)
        
    
    # Note that we need to compute by dataset instead of by batch
    train_preds = torch.vstack(train_pairs['preds'])
    train_labels = torch.vstack(train_pairs['labels'])
    
    train_acc = compute_accuracy(preds = train_preds, 
                                    labels = train_labels)
    
    model.eval()
    with torch.no_grad():
        for id, batch in enumerate(val_loader):
            inputs = batch['input_ids'].to(device)
            attnmasks = batch['attention_mask'].to(device)
            labels = batch['label'].to(device)
            logits = model(inputs, attnmasks).logits
            # storing results 
            preds = F.softmax(logits, dim = 1)
            loss = criterion(preds, labels)
            
            preds = preds.detach().cpu().numpy()
            labels = labels.detach().cpu().numpy()
            val_pairs['labels'].append(labels)
            val_pairs['preds'].append(preds)
            epoch_loss['val'] += loss.item()
        
        # Note that we need to compute by dataset instead of by batch
        val_preds = np.vstack(val_pairs['preds'])
        val_labels = np.vstack(val_pairs['labels'])
        val_acc = compute_accuracy(labels = val_labels, 
                                       preds = val_preds)
    wandb.log({
        'train macro f1': train_acc['macro f1'],
        'train weighted f1': train_acc['weighted f1'],
        'train loss': epoch_loss['train']/len(train_loader),
        'val macro f1': val_acc['macro f1'],
        'val weighted f1':val_acc['weighted f1'],
        'val loss': epoch_loss['val']/len(val_loader)
        })
    print(f"[{epoch+1+base}/{extra_epochs+base}] |TRAIN loss: {epoch_loss['train']/len(train_loader):.3f} |VAL loss:{epoch_loss['val']/len(val_loader):.3f},\
          \n         VAL F1 macro: {val_acc['macro f1']:.3f}, weighted: {val_acc['weighted f1']:.3f}")
    if val_acc['macro f1'] > max_val_acc:
        torch.save(model, checkpoint_path)
        print('🏆 Saving model!')
        max_val_acc = val_acc['macro f1']
        best_val_pairs  = val_pairs
        print('===================================')

## Classification Report 

from sklearn.metrics import classification_report
cat2idx = joblib.load(f'{datadir}/category/cat2idx.pkl')
labels = sorted(cat2idx.items(), key = lambda x:x[1])
keys = [x[0] for x in labels]
preds = np.where(np.vstack(best_val_pairs['preds']) <= 0.5, 0, 1)
labels = np.vstack(best_val_pairs['labels'])
ZD = 1 
print(classification_report(labels, 
                            preds, target_names= keys, zero_division = ZD))
