==> codes/category/40.0_add_valid_string.ipynb <==
## 40.0 Post-hoc

Author: Nana
Date: 2022.5.15
1. run inference on valid set, save them 
2. identify those categories with low f1 score
3. run fasttext keywords string exact matching
    if exact match exists and the probability of the selected category is below a threshold h,
    then replace the model-prediction with fasttext choice. 
    

import joblib, sys
import numpy as np
import pandas as pd
import random
import torch 
import os
import torch.nn as nn
import torch.nn.functional as F
checkpoint_path = './checkpoints/0510-0939/model.ckpt'
dataset_path = './dataset/encoded_dataset_30.0'

import datasets
from datasets import Dataset, load_metric

dataset = datasets.load_from_disk(dataset_path)
valset = dataset['test']

print(valset[0].keys())

# load the fine-tuned model for inference
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
print(device)

model = torch.load(checkpoint_path)
model.to(device)
print()

!nvidia-smi

MODELNAME = 'bert-base-chinese'
from transformers import BertTokenizerFast
tokenizer = BertTokenizerFast.from_pretrained(MODELNAME)
def collate_fn(batch):
    keys = batch[0].keys()
    # print(keys)
    # 'id', 'name', 'label', 'seg_name', 'label_name', 'input_ids', 'token_type_ids', 'attention_mask', 'labels'
    to_tensor = ['input_ids', 'token_type_ids', 'attention_mask']
    new_batch = {k:[d[k] for d in batch] for k in keys if k in to_tensor}
    new_batch = tokenizer.pad(
        new_batch,
        padding= "max_length",
        max_length= 100,
        return_tensors="pt",
    )
    label = [d['label'] for d in batch]
    ids = [d['id'] for d in batch]
    new_batch['label'] = torch.tensor(label, dtype=torch.float32)
    new_batch["id"] = torch.tensor(ids, dtype=torch.int64)
    return new_batch

from torch.utils.data import DataLoader
batch_size = 20
val_loader = DataLoader(valset, 
                        batch_size = batch_size, 
                        collate_fn=collate_fn, 
                        shuffle = False)

from sklearn.metrics import f1_score, classification_report
def compute_accuracy(labels, preds, threshold = 0.5, zd = 1):
    preds = np.where(preds <= threshold, 0, 1)
    return {'macro f1': f1_score(labels, preds, average = 'macro', zero_division = zd), 
            'weighted f1': f1_score(labels, preds, average = 'weighted', zero_division = zd)}

model.to(device)
model.eval()
with torch.no_grad():
    val_pairs = {'preds': [], 
                'labels': []}
    for id, batch in enumerate(val_loader):
        inputs = batch['input_ids'].to(device)
        attnmasks = batch['attention_mask'].to(device)
        labels = batch['label'].to(device)
        logits = model(inputs, attnmasks).logits
        # storing results 
        preds = F.softmax(logits, dim = 1)

        preds = preds.detach().cpu().numpy()
        labels = labels.detach().cpu().numpy()
        val_pairs['labels'].append(labels)
        val_pairs['preds'].append(preds)

    # Note that we need to compute by dataset instead of by batch
    val_preds = np.vstack(val_pairs['preds'])
    val_labels = np.vstack(val_pairs['labels'])
    val_acc = compute_accuracy(labels = val_labels, 
                                   preds = val_preds)
    raw_val_preds = val_preds
    validation_predictions = np.where(val_preds <= 0.5, 0, 1)
    
    print("🍎 Finished inferencing.")
    print(f"[info]|VAL F1 macro: {val_acc['macro f1']:.3f}, weighted: {val_acc['weighted f1']:.3f}")
    

len(validation_predictions)==len(valset)

predpath = './predictions/0.63_val_predictions.pkl'
labelspath = './predictions/encoded_30.0_labels.pkl'
probspath = './predictions//0.63_val_predictions_probs.pkl'
# joblib.dump(validation_predictions, predpath)
# joblib.dump(val_labels, labelspath)
# joblib.dump(raw_val_preds, probspath)
valpreds = joblib.load(predpath)
vallabels = joblib.load(labelspath)


### Run afterwards

# hyperparams
N = 217 # category數量

## f1_score(y_true, y_pred, average=None)
from sklearn.metrics import precision_recall_fscore_support as score
_, _,f1byclass, support = score(y_true = vallabels, y_pred = valpreds, zero_division = 1,  # no testing sample的class其f1取1
                     average = None)

# !conda install scikit-learn --yes

cat2idx = joblib.load(f'./category/cat2idx.pkl')
idx2cat = {v:k for k, v in cat2idx.items()}

poors = []
for i in range(N):
    if f1byclass[i] < 0.1:
        poors.append((idx2cat[i], f1byclass[i], support[i]))

poors # 71類 (category, category f1, support)
# support: number of this category in validation set 

### Add 23.0 Attempt 1: String Exact Match

## load fasttext model 
from gensim.models import FastText
fmodelpath = f'./fasttext/fasttext.model'
fm = FastText.load(fmodelpath) 

#!conda install gensim==3.8.3
import gensim
# downgrade the hecking gensim for model loading compatibility 

print(gensim.__version__)

## get category keywords dictionary 
TOPN = 10
from collections import defaultdict
cat2keywords= defaultdict(list)
for cat in cat2idx.keys():
    if cat in fm.wv:
        cat2keywords[cat] = fm.wv.most_similar(cat, topn = TOPN)
        for subcat in cat.split('/'):
            # 有些會有 快煮麵/乾拌麵 形式
            cat2keywords[cat].append((subcat, 1))
    else:
        tokens = cat2keywords[cat]
        for token in tokens:
            cat2keywords.extend(fm.wv.most_similar(token, topn = TOPN))

for (poorclass, f1, sup) in poors:
    print(poorclass)
    print(cat2keywords[poorclass])

poorpath = './category/poor_classf1.pkl'
joblib.dump(poors, poorpath)
cat2keywordspath = './category/cat2kw.pkl'
joblib.dump(cat2keywords, cat2keywordspath)

cat2keywordspath = './category/cat2kw.pkl'
cat2keywords = joblib.load(cat2keywordspath)
cat2keywords['快煮麵/乾拌麵']

'''
for i, sample in enumerate(samples):
    hit_vector = np.zeros(N)
    name = sample['name']
    for cat, pair in cat2kw.items():
        ph_list, ph_probs = pair
        pat = re.compile(r'(?:{})'.format('|'.join(map(re.escape, ph_list))))
        found = re.findall(pat, name)
        cat_hit_probs = sum(ph_probs[x] for x in found)
        catid = cat2idx[cat]
        hit_vector[catid] = cat_hit_probs
    samples[i]['hitsprobs'] = hit_vector 
''' 

poors_ = [x[0] for x in poors]

poor_cat2kw = {k:v for k, v in cat2keywords.items() if k in poors_}

def getclass(vec, THR):
    classes = set()
    pred = np.where(vec >= THR)[0].tolist()
    for cls in pred:
        classes.add(idx2cat[cls])
    return classes

valprobs = joblib.load(probspath)

#### string matching...

import re
THR = 0.99
# use valprobs
changed = 0
val_names = valset['name']


for i, name in enumerate(val_names):
    # pred vector is the prediction proabability vector of i-th validation data 
    pred_vector = valprobs[i, :]
    
    # for the classes that have poor f1 score: 
    for cat, pairs in poor_cat2kw.items():
        kws = [x[0] for x in pairs]
        kwprobs = [x[1] for x in pairs]
        kwdict = {kw:kwprob for kw, kwprob in zip(kws, kwprobs)}
        # use regex to find if the keyword appears in validation data's `name` 
        pat = re.compile(r'(?:{})'.format('|'.join(map(re.escape, kws))))
        # `hits` means the number of unique keywords found in 'name'
        hits = re.findall(pat, name)
        # for those "hit keywords", add up their cos_similarity
        # I personally call the summed cos_sim `cat_hit_probs` (category hit probability)
        # which is not completely accurate so bear with it 
        cat_hit_probs = sum(kwdict[hit] for hit in hits)
        # find the class' corresponding index
        catid = cat2idx[cat]
        # !!!!!!!!!!!!!!!!!!
        # !!! revise the prediction vector by plugging for this class the `cat_hit_probs`!!!
        pred_vector[catid] = cat_hit_probs

idx2cat[np.argmax(valprobs[0])] # 預測

idx2cat[np.argmax(vallabels[0])] # 實際標記

# number of predictions changed: 
changed

## 🏆 probability threshold = 0.5
當初訓練的門檻也是調0.5，macro f1可至0.71左右

### np.where的threshold設0.5： 
compute_accuracy(vallabels, valprobs, threshold = 0.5)

比較：$ BERT訓練成果
🍎 VAL F1 macro: 0.629, weighted: 0.856

labels = sorted(cat2idx.items(), key = lambda x:x[1])
keys = [x[0] for x in labels]

revised_preds = np.where(valprobs <= 0.5, 0, 1)
print(classification_report(y_true=vallabels, 
                            y_pred = revised_preds, 
                            zero_division = 1, 
                            target_names=keys))

## 🏆 BEST probability threshold = 0.995(greedy search) 

目前看起來最佳的threshold如上，macro f1可至0.76左右。

簡單來說這個演算法流程就是**訓練多標籤BERT（macro f1=0.639），以其預測結果輔以fasttext所訓練出的top 10關鍵字matching（macro f1 = 0.738）**。
都是以資料集本身餵入處理與切割train test，未經過class數量的篩選與人為操作，如test set分布和valid set相似，可預估最後表現應該也是落在**0.63 ~0.74**之間。

🍀 我這邊應該就做到這一步，接下來我會開始把這些步驟串在一起寫inference script。

THR = 0.85
best_THR = 0
maxi = 0
while THR < 1:
    THR += 1e-3
    res = compute_accuracy(vallabels, valprobs, threshold = THR)
    if res['macro f1'] > maxi:
        maxi = res['macro f1']
        best_THR = THR

print(best_THR, maxi)

revised_preds = np.where(valprobs <= best_THR, 0, 1)
print(classification_report(y_true=vallabels, 
                            y_pred = revised_preds, 
                            zero_division = 1, 
                            target_names=keys))

## Make prediction dataframes (csv)

revised_preds.shape

valsetsize = revised_preds.shape[0]
preds = []
labels = []
for i in range(valsetsize):
    pred = revised_preds[i, :]
    predicted_classes = np.where(pred==1)[0]
    product = ';'.join([idx2cat[pclass] for pclass in predicted_classes])
    preds.append(product)
colsdict = {'name': [valset[i]['name'] for i in range(valsetsize)],
           'predictions': preds,
           'labels': [';'.join(valset[i]['label_name']) for i in range(valsetsize)]}
df = pd.DataFrame.from_dict(colsdict)

df.head()

df.to_csv('./predictions/0.746_category_val_parallel.csv')
